Organizando toda a estrutura do código
