+++ src/pages/Layout/Scss/ReservaPedido.module.scss
@@ -3,17 +3,16 @@
     object-position: center;
 }
 .text {
-    padding: 3em 1em;
+    margin: 3em 1em;
 }
 .btnRD {
-    margin: 2em 0;
+    margin-bottom: 2em;
     color: #fff;
     background-color: #997d34d1;
-    padding: 10px 15px;
+    padding: 10px;
     border-radius: 10px;
     text-transform: uppercase;
     font-size: 1.1rem;
-    cursor: pointer;
 }
 .espaco {
     margin-bottom: 2em;
