+++ src/pages/Layout/_header.tsx
@@ -1,10 +1,10 @@
 import Image from 'next/image'
 import Link from 'next/link'
 
-import logoHeader from '../../../../public/logo3.webp'
-import UserLoginButton from '../LayoutLogin/_userLoginButton'
+import logoHeader from '../../../public/logo3.webp'
+import UserLoginButton from './_userLoginButton'
 
-import styles from '../ScssMain/Main.module.scss'
+import styles from './Scss/Main.module.scss'
 export default function Header() {
   return (
     <header className={styles.Header}>
