+++ src/pages/cardapio/Layout/_bebidas.tsx
@@ -10,13 +10,13 @@ import SaladaRussa from '../../../../public/saladarussacamarao.webp'
 import tabuaEnchidos from '../../../../public/tabua-de-enchidos.webp'
 import tabuaQueijos from '../../../../public/tabua-de-queijos.webp'
 
-import styles from '../ScssCardapio/Cardapio.module.scss'
+import styles from './Scss/Cardapio.module.scss'
 
-export default function Pratos() {
+export default function Bebidas() {
   return (
     <>
       <div>
-        <h2 className={styles.PageName}>Pratos</h2>
+        <h2 className={styles.PageName}>Bebidas</h2>
       </div>
       <div className={styles.CDContainer}>
         <CardapioItens1OP
@@ -26,6 +26,38 @@ export default function Pratos() {
           }
           Preco={'19,90'}
           imgPrato={Canapes}
+        />
+        <CardapioItens1OP
+          NomeDoPrato={'Canapés'}
+          Desc={
+            ' 2x pastas de barrar, 1x mostarda, 1x ketchup, 1x maionese, 2x cebolas, 2x tomates cherry, 2x ovos cozidos, 2x azeitonas 2x queijo Gorgonzola.'
+          }
+          Preco={'19,90'}
+          imgPrato={SaladaRussa}
+        />
+        <CardapioItens1OP
+          NomeDoPrato={'Canapés'}
+          Desc={
+            ' 2x pastas de barrar, 1x mostarda, 1x ketchup, 1x maionese, 2x cebolas, 2x tomates cherry, 2x ovos cozidos, 2x azeitonas 2x queijo Gorgonzola.'
+          }
+          Preco={'19,90'}
+          imgPrato={SaladaPolvo}
+        />
+        <CardapioItens1OP
+          NomeDoPrato={'Canapés'}
+          Desc={
+            ' 2x pastas de barrar, 1x mostarda, 1x ketchup, 1x maionese, 2x cebolas, 2x tomates cherry, 2x ovos cozidos, 2x azeitonas 2x queijo Gorgonzola.'
+          }
+          Preco={'19,90'}
+          imgPrato={SaladaAtum}
+        />
+        <CardapioItens1OP
+          NomeDoPrato={'Canapés'}
+          Desc={
+            ' 2x pastas de barrar, 1x mostarda, 1x ketchup, 1x maionese, 2x cebolas, 2x tomates cherry, 2x ovos cozidos, 2x azeitonas 2x queijo Gorgonzola.'
+          }
+          Preco={'19,90'}
+          imgPrato={Marmita}
         />
 
         <CardapioItens1OP
@@ -54,7 +86,29 @@ export default function Pratos() {
           PrecoP={'30,00'}
           PrecoM={'40,00'}
           PrecoG={'50,00'}
-          imgPrato={Marmita}
+          imgPrato={SaladaOrelha}
+        />
+
+        <CardapioItensSize
+          NomeDoPrato={'teste'}
+          Desc={
+            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
+          }
+          PrecoP={'30,00'}
+          PrecoM={'40,00'}
+          PrecoG={'50,00'}
+          imgPrato={SaladaOrelha}
+        />
+
+        <CardapioItensSize
+          NomeDoPrato={'teste'}
+          Desc={
+            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
+          }
+          PrecoP={'30,00'}
+          PrecoM={'40,00'}
+          PrecoG={'50,00'}
+          imgPrato={SaladaOrelha}
         />
 
         <CardapioItensSize
@@ -98,29 +152,7 @@ export default function Pratos() {
           PrecoP={'30,00'}
           PrecoM={'40,00'}
           PrecoG={'50,00'}
-          imgPrato={SaladaAtum}
-        />
-
-        <CardapioItensSize
-          NomeDoPrato={'teste'}
-          Desc={
-            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
-          }
-          PrecoP={'30,00'}
-          PrecoM={'40,00'}
-          PrecoG={'50,00'}
-          imgPrato={SaladaPolvo}
-        />
-
-        <CardapioItensSize
-          NomeDoPrato={'teste'}
-          Desc={
-            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
-          }
-          PrecoP={'30,00'}
-          PrecoM={'40,00'}
-          PrecoG={'50,00'}
-          imgPrato={SaladaRussa}
+          imgPrato={SaladaOrelha}
         />
       </div>
     </>
