+++ src/pages/cardapio/Layout/_cardapioItens1OP.tsx
@@ -1,5 +1,5 @@
 import Image, { StaticImageData } from 'next/image'
-import styles from '../ScssCardapio/Cardapio.module.scss'
+import styles from './Scss/Cardapio.module.scss'
 
 interface CardapioInterface {
   NomeDoPrato: string
