+++ src/pages/cardapio/Layout/_navBarCardapio.tsx
@@ -1,6 +1,6 @@
 import { FaArrowLeft } from 'react-icons/fa'
 import Link from 'next/link'
-import styles from '../ScssCardapio/Cardapio.module.scss'
+import styles from './Scss/Cardapio.module.scss'
 
 export default function NavBarCardapio() {
   return (
