+++ src/pages/cardapio/Layout/_porcao.tsx
@@ -1,6 +1,3 @@
-import CardapioItens1OP from './_cardapioItens1OP'
-import CardapioItensSize from './_cardapioItensSize'
-
 import Canapes from '../../../../public/canapes.webp'
 import Marmita from '../../../../public/marmita.webp'
 import SaladaAtum from '../../../../public/saladaatum.webp'
@@ -10,13 +7,16 @@ import SaladaRussa from '../../../../public/saladarussacamarao.webp'
 import tabuaEnchidos from '../../../../public/tabua-de-enchidos.webp'
 import tabuaQueijos from '../../../../public/tabua-de-queijos.webp'
 
-import styles from '../ScssCardapio/Cardapio.module.scss'
+import CardapioItens1OP from './_cardapioItens1OP'
+import CardapioItensSize from './_cardapioItensSize'
 
-export default function Bebidas() {
+import styles from './Scss/Cardapio.module.scss'
+
+export default function Procao() {
   return (
     <>
       <div>
-        <h2 className={styles.PageName}>Bebidas</h2>
+        <h2 className={styles.PageName}>Proção</h2>
       </div>
       <div className={styles.CDContainer}>
         <CardapioItens1OP
@@ -26,38 +26,6 @@ export default function Bebidas() {
           }
           Preco={'19,90'}
           imgPrato={Canapes}
-        />
-        <CardapioItens1OP
-          NomeDoPrato={'Canapés'}
-          Desc={
-            ' 2x pastas de barrar, 1x mostarda, 1x ketchup, 1x maionese, 2x cebolas, 2x tomates cherry, 2x ovos cozidos, 2x azeitonas 2x queijo Gorgonzola.'
-          }
-          Preco={'19,90'}
-          imgPrato={SaladaRussa}
-        />
-        <CardapioItens1OP
-          NomeDoPrato={'Canapés'}
-          Desc={
-            ' 2x pastas de barrar, 1x mostarda, 1x ketchup, 1x maionese, 2x cebolas, 2x tomates cherry, 2x ovos cozidos, 2x azeitonas 2x queijo Gorgonzola.'
-          }
-          Preco={'19,90'}
-          imgPrato={SaladaPolvo}
-        />
-        <CardapioItens1OP
-          NomeDoPrato={'Canapés'}
-          Desc={
-            ' 2x pastas de barrar, 1x mostarda, 1x ketchup, 1x maionese, 2x cebolas, 2x tomates cherry, 2x ovos cozidos, 2x azeitonas 2x queijo Gorgonzola.'
-          }
-          Preco={'19,90'}
-          imgPrato={SaladaAtum}
-        />
-        <CardapioItens1OP
-          NomeDoPrato={'Canapés'}
-          Desc={
-            ' 2x pastas de barrar, 1x mostarda, 1x ketchup, 1x maionese, 2x cebolas, 2x tomates cherry, 2x ovos cozidos, 2x azeitonas 2x queijo Gorgonzola.'
-          }
-          Preco={'19,90'}
-          imgPrato={Marmita}
         />
 
         <CardapioItens1OP
@@ -75,6 +43,61 @@ export default function Bebidas() {
             '3x queijo de vaca, 3x queijo Camembert, 2x queijo Cantal, 2x queijo Comté, 2x queijo Gorgonzola'
           }
           Preco={'19,90'}
+          imgPrato={SaladaAtum}
+        />
+
+        <CardapioItensSize
+          NomeDoPrato={'teste'}
+          Desc={
+            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
+          }
+          PrecoP={'30,00'}
+          PrecoM={'40,00'}
+          PrecoG={'50,00'}
+          imgPrato={SaladaOrelha}
+        />
+
+        <CardapioItensSize
+          NomeDoPrato={'teste'}
+          Desc={
+            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
+          }
+          PrecoP={'30,00'}
+          PrecoM={'40,00'}
+          PrecoG={'50,00'}
+          imgPrato={Marmita}
+        />
+
+        <CardapioItensSize
+          NomeDoPrato={'teste'}
+          Desc={
+            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
+          }
+          PrecoP={'30,00'}
+          PrecoM={'40,00'}
+          PrecoG={'50,00'}
+          imgPrato={SaladaPolvo}
+        />
+
+        <CardapioItensSize
+          NomeDoPrato={'teste'}
+          Desc={
+            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
+          }
+          PrecoP={'30,00'}
+          PrecoM={'40,00'}
+          PrecoG={'50,00'}
+          imgPrato={SaladaOrelha}
+        />
+
+        <CardapioItensSize
+          NomeDoPrato={'teste'}
+          Desc={
+            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
+          }
+          PrecoP={'30,00'}
+          PrecoM={'40,00'}
+          PrecoG={'50,00'}
           imgPrato={tabuaQueijos}
         />
 
@@ -86,62 +109,7 @@ export default function Bebidas() {
           PrecoP={'30,00'}
           PrecoM={'40,00'}
           PrecoG={'50,00'}
-          imgPrato={SaladaOrelha}
-        />
-
-        <CardapioItensSize
-          NomeDoPrato={'teste'}
-          Desc={
-            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
-          }
-          PrecoP={'30,00'}
-          PrecoM={'40,00'}
-          PrecoG={'50,00'}
-          imgPrato={SaladaOrelha}
-        />
-
-        <CardapioItensSize
-          NomeDoPrato={'teste'}
-          Desc={
-            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
-          }
-          PrecoP={'30,00'}
-          PrecoM={'40,00'}
-          PrecoG={'50,00'}
-          imgPrato={SaladaOrelha}
-        />
-
-        <CardapioItensSize
-          NomeDoPrato={'teste'}
-          Desc={
-            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
-          }
-          PrecoP={'30,00'}
-          PrecoM={'40,00'}
-          PrecoG={'50,00'}
-          imgPrato={SaladaOrelha}
-        />
-
-        <CardapioItensSize
-          NomeDoPrato={'teste'}
-          Desc={
-            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
-          }
-          PrecoP={'30,00'}
-          PrecoM={'40,00'}
-          PrecoG={'50,00'}
-          imgPrato={SaladaOrelha}
-        />
-
-        <CardapioItensSize
-          NomeDoPrato={'teste'}
-          Desc={
-            'Lorem ipsum dolor sit amet consectetur, adipisicing elit. Repellat nulla minus'
-          }
-          PrecoP={'30,00'}
-          PrecoM={'40,00'}
-          PrecoG={'50,00'}
-          imgPrato={SaladaOrelha}
+          imgPrato={SaladaRussa}
         />
 
         <CardapioItensSize
